implement like and unlike book

+++ server/BookReaderAPI/BookReaderAPI/DTOs/BookDTO.cs
@@ -9,8 +9,8 @@
         public string? Description { get; set; }
         public int? CoverImgFileId { get; set; }
         public Int64? Views { get; set; }
-        public int Likes { get; set; } = 0;
-        public int Comments { get; set; } = 0;
+        public Int64 Likes { get; set; } = 0;
+        public Int64 Comments { get; set; } = 0;
         public DateTime? CreatedAt { get; set; }
         public DateTime? UpdatedAt { get; set; }
     }
